requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-dotenv
aiogram
python-jose[cryptography]
bcrypt
feedparser
scikit-learn
apscheduler
